KimHanBin-dev/backend/requirements.txt
bcrypt==3.2.0
blinker==1.4
cachelib==0.3.0
cffi==1.14.6
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
Deprecated==1.2.12
Flask==1.1.4
Flask-Bcrypt==0.7.1
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Mail==0.9.1
flask-praetorian==1.2.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
greenlet==1.1.1
importlib-metadata==4.6.4
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
passlib==1.7.4
pendulum==2.1.2
py-buzz==1.0.3
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytzdata==2020.1
six==1.16.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.5.0
PyMySQL==1.0.2
mysql-connector-python==8.0.26